Make expansion breadth-first and multi-phase
Breadth-first: avoids repeated evaluation of nodes near the root.

Multi-phase: in the end I could think of no simpler way to cope with the
requirements of both data templating and querying: the data templating
happens in the first phase, and the querying can then take place on the
expanded data. This is implemented by allowing a numeric suffix in the
“.ruth” file infix. “.ruth” is the same as “.ruth0”. Templated files are
sorted by their suffix, so that all “.ruth0” files are expanded before all
“.ruth1” files etc.

Extend the data-templating test to 3-level templating, and make it
multi-phase.