Remove repeated iteration
This is not necessary now that we have the “serialize” function.

Rejig the ruth.xq XQuery functions to re-evaluate their results where
necessary.

Hence, remove --max-iterations and the non-termination test (which tested
the error condition of exceeding the maximum number of iterations).

The workaround to namespace the global variables does not work, so remove
it. The obvious problem is that the variable will only be set when the
built-in ruth.xq is registered. Actually, though something odder seems to be
happening: some other value of the referenced global is being used. So,
refer directly to the non-namespaced globals, and prefix their names with
“ruth_” to avoid clashes.