Add support for XQuery Update Facility
Make it possible for a file to have both a `.in` and a `.ruth` suffix, as
this is needed for updates to be useful. Allow both the suffixes to occur
before the penultimate position.

Include a test and documentation for updates.

One or two other minor improvements to the code.