Turn error message into assertion
The only way that a file to be copied can be not a file is if the filing
system has been modified since the file was originally read (a race
condition).

Fix test coverage by turning an error that is never executed by the tests
into an assertion. The assertion is needed for TypeScript to infer the type
correctly, but of course the race condition remains, and fs.copyFileSync
would fail in that case.