Factor out XmlDir class
Add new exported class XmlDir. This makes it possible to just read an input
tree and have it returned as an XML object (or an XML document, via the
formatXML method).

Use the platform os.EOL as the default lineSeparator argument to formatXML.

Do not do expensive computation of debug output (i.e. call formatXML) unless
we are actually debugging. As a result, set DEBUG to ‘*’ rather than ‘yes’
for the tests, to ensure full coverage.

Add a new helper function for the tests, assertStringEqualToFile, which we
use for the new test of XMLDir.