Revert "Remove custom ruth:map function"
This reverts commit 6b4b546e6385934231fee681694275c6edec51a8.

There really is no other sensible way to transform parts of the input. One
would have to write a function that recurses over whatever is to be
transformed, which is a lot of unnecessary work. XSLT would be better
suited, but again, using an entire other language just for this task seems
overkill.